Additions to the Website
I finished making the wireframe and then I started on the official creation of the website interface. I got the website to a point that I'm going to add finishing touches and then consider the website done. I added two points where Javascript is used: the picture and picture description navigator, and the submit response when the form is submitted. The picture and picture description navigator makes it where when you click the "next" button, it will give a new image along with a new image description. There are a max of four pictures stored in the picture navigator. The final use of Javascript was with the submit function at the bottom of the website. When the user clicks "submit", the user's input is deleted and the form is hidden. A new message then replaces the form.